+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - refactor: enchance dns resolver to update upstreams.
 - refactor: optimize the information reported by OpenTelemetry tracing and redesign config.
-- refactor: upgrade and rename access log directives
+- [[#7](https://github.com/nite-coder/bifrost/pull/7)]: upgrade and rename access log directives
+- [[#8](https://github.com/nite-coder/bifrost/pull/8)]: new response-transformer middleware
 
 ## 2024-12-22 0.1.0